requirements.txt
certifi==2017.4.17
chardet==3.0.4
future==0.16.0
idna==2.5
oauthlib==2.0.2
python-twitter==3.3
requests==2.18.1
requests-oauthlib==0.8.0
urllib3==1.21.1
